+++ src/store/actions/search-conv.ts
@@ -25,7 +25,7 @@ export type SearchConvReturn = {
 
 export const searchConv = createAsyncThunk<SearchConvReturn, SearchConvParameters>(
 	'conversations/searchConv',
-	async ({ conversationId, folderId, fetch = '0' }) => {
+	async ({ conversationId, folderId, fetch = 'all' }) => {
 		const result = (await soapFetch<SearchConvRequest, SearchConvResponse>('SearchConv', {
 			_jsns: 'urn:zimbraMail',
 			cid: conversationId,
@@ -39,7 +39,7 @@ export const searchConv = createAsyncThunk<SearchConvReturn, SearchConvParameter
 			html: 1,
 			max: 250000
 		})) as SearchConvResponse;
-		const messages = map(result?.m ?? [], (msg) => normalizeMailMessageFromSoap(msg, false));
+		const messages = map(result?.m ?? [], (msg) => normalizeMailMessageFromSoap(msg, true));
 
 		return {
 			messages,