+++ src/views/settings/displaying-messages-settings.jsx
@@ -10,8 +10,7 @@ import {
 	FormSubSection,
 	Select,
 	Input,
-	Checkbox,
-	Padding
+	Checkbox
 } from '@zextras/carbonio-design-system';
 import Heading from './components/settings-heading';
 import {
@@ -21,6 +20,7 @@ import {
 	findLabel,
 	ConversationSortingSettings
 } from './components/utils';
+import { displayingMessagesSubSection } from './subsections';
 
 export default function DisplayingMessagesSettings({ t, settingsObj, updateSettings }) {
 	const checkNewMailOptions = useMemo(
@@ -35,12 +35,10 @@ export default function DisplayingMessagesSettings({ t, settingsObj, updateSetti
 	const displayMailOptions = useMemo(() => DisplayMailOptions(t), [t]);
 	const messageSelectionOptions = useMemo(() => MessageSelectionOptions(t), [t]);
 	const conversationSortingSettings = useMemo(() => ConversationSortingSettings(t), [t]);
+	const sectionTitle = useMemo(() => displayingMessagesSubSection(t), [t]);
 
 	return (
-		<FormSubSection
-			label={t('settings.label.display_messages', 'Displaying Messages')}
-			padding={{ all: 'medium' }}
-		>
+		<FormSubSection id={sectionTitle.id} label={sectionTitle.label} padding={{ all: 'medium' }}>
 			<Row width="100%" padding={{ horizontal: 'small', vertical: 'small' }}>
 				<Select
 					label={t('settings.label.conversation_ordering', 'Conversation ordering')}
@@ -73,7 +71,9 @@ export default function DisplayingMessagesSettings({ t, settingsObj, updateSetti
 					label={t('settings.label.display_mail', 'Display mail')}
 					items={displayMailOptions}
 					onChange={(view) =>
-						updateSettings({ target: { name: 'zimbraPrefMessageViewHtmlPreferred', value: view } })
+						updateSettings({
+							target: { name: 'zimbraPrefMessageViewHtmlPreferred', value: view }
+						})
 					}
 					defaultSelection={{
 						label: findLabel(displayMailOptions, settingsObj.zimbraPrefMessageViewHtmlPreferred),